+++ web-frontend/src/main/java/org/followfa/postings/query/QueryPostingsService.java
@@ -3,5 +3,5 @@ package org.followfa.postings.query;
 import java.util.List;
 
 public interface QueryPostingsService {
-	List<? extends QueriedPosting> listPostingsForCurrentUser(long userId, long maxResults);
+	List<? extends QueriedPosting> listPostingsForCurrentUser(long userId, int maxResults);
 }
