+++ web-frontend/src/main/java/org/followfa/web/posting/PostingTranslationService.java
@@ -5,5 +5,5 @@ import java.util.stream.Stream;
 public interface PostingTranslationService {
 	void createNewPosting(long userId, String postingText);
 
-	Stream<PostingViewModel> getNewestPostingsForCurrentUser(long userId, long maxResults);
+	Stream<PostingViewModel> getNewestPostingsForCurrentUser(long userId, int maxResults);
 }
